Update Seeder.java (#170)
* Update Seeder.java

* Update RacunSeeder.java

* Update MailTest.java

+++ user-service/src/main/java/rs/edu/raf/repository/Seeder.java
@@ -14,10 +14,13 @@ public class Seeder implements CommandLineRunner {
 
     KorisnikRepository korisnikRepository;
     RadnikRepository radnikRepository;
+    boolean reseed = false;
 
     @Override
     public void run(String... args) throws Exception {
         try{
+            if(!reseed)
+                return;
             korisnikRepository.deleteAll();
             
             Korisnik korisnik = new Korisnik();
